unified fixture
